pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["marcbeil <marcbeil2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.5"
sqlite-vss = "^0.1.2"
langchain-openai = "^0.1.8"
langchain-core = "^0.2.9"
sqlean = "^0.0.3"
openai = "^1.35.3"
faiss-cpu = "^1.8.0.post1"
dbml-sqlite = "^0.3.3"
sqlalchemy = "^2.0.31"
flask = "^3.0.3"
pandas = "^2.2.2"
dash-bootstrap-components = "^1.6.0"
ratelimit = "^2.2.1"
db = "^0.1.1"
scikit-learn = "^1.5.1"
python-levenshtein = "^0.25.1"
roman = "^4.2"
flask-cors = "^4.0.1"
matplotlib = "^3.9.2"
ace-tools = "^0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
